# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: D&D Text Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/{characterId}/inventory:
    get:
      operationId: getInventory
      tags: [players]
      summary: Get character inventory
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: addInventoryItem
      tags: [players]
      summary: Add item to inventory
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddInventoryItemRequest"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /characters/{characterId}/inventory/{itemId}:
    delete:
      operationId: removeInventoryItem
      tags: [players]
      summary: Remove/use item from inventory
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /characters/{characterId}/achievements:
    get:
      operationId: getAchievements
      tags: [players]
      summary: Get character achievements
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"
...
  /campaigns/{campaignId}/join:
    post:
      operationId: joinCampaign
      tags: [campaigns]
      summary: Join a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinCampaignRequest"
      responses:
        "200":
          description: Joined campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
...
  /sessions/{sessionId}/action:
    post:
      operationId: performAction
      tags: [game]
      summary: Perform a game action (AI DM responds)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionRequest"
      responses:
        "200":
          description: DM response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameActionResponse"
  /sessions/{sessionId}/history:
    get:
      operationId: getSessionHistory
      tags: [game]
      summary: Get session message history
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameMessage"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        characterId:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
        quantity:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, characterId, name, type, quantity, createdAt]
    AddInventoryItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        quantity:
          type: integer
      required: [name, type, quantity]
    Achievement:
      type: object
      properties:
        id:
          type: integer
        characterId:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        unlockedAt:
          type: string
          format: date-time
      required: [id, characterId, title, description, icon, unlockedAt]
...
    JoinCampaignRequest:
      type: object
      properties:
        playerId:
          type: integer
        characterId:
          type: integer
        inviteCode:
          type: string
      required: [playerId, characterId]
    GameSession:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        playerId:
          type: integer
        characterId:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, campaignId, playerId, characterId, status, createdAt]
...
    GameActionRequest:
      type: object
      properties:
        action:
          type: string
        characterId:
          type: integer
      required: [action]
    GameActionResponse:
      type: object
      properties:
        narrative:
          type: string
        sessionId:
          type: integer
        xpGained:
          type: integer
        hpChange:
          type: integer
        diceRequest:
          type: string
        isDead:
          type: boolean
        newAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        newItems:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        newLocation:
          type: string
      required: [narrative, sessionId]
    GameMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, sessionId, role, content, createdAt]
...