# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Text Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: number;
  characterId: number;
  title: string;
  description: string;
  icon: string;
  unlockedAt: Date;
}