# lib/api-zod/src/generated/types/addInventoryItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Text Game API
 * OpenAPI spec version: 0.1.0
 */

export interface AddInventoryItemRequest {
  name: string;
  description?: string;
  type: string;
  quantity: number;
}