# lib/api-zod/src/generated/types/gameActionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Text Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Achievement } from "./achievement";
import type { InventoryItem } from "./inventoryItem";

export interface GameActionResponse {
  narrative: string;
  sessionId: number;
  xpGained?: number;
  hpChange?: number;
  diceRequest?: string;
  isDead?: boolean;
  newAchievements?: Achievement[];
  newItems?: InventoryItem[];
  newLocation?: string;
}